kneighborsclassifier_grpc/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
protobuf==3.20.0
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
